refactor: split up correctors for readability
It's much easier to reason about, if you don't feel like you have to
take all that other code into account. But actually you don't have to.

(cherry picked from commit 06c9338e100679f0bfa282df762a13e2e6d0b6b6)
